Convert quickstart to typescript
